+++ unilife-server/src/main/java/com/unilife/model/entity/Schedule.java
@@ -62,7 +62,7 @@ public class Schedule implements Serializable {
     /**
      * 提醒时间（分钟）
      */
-    private Byte reminder;
+    private Integer reminder;
 
     /**
      * 显示颜色
