+++ unilife-server/src/main/java/com/unilife/model/vo/ScheduleVO.java
@@ -58,7 +58,7 @@ public class ScheduleVO {
     /**
      * 提醒时间（分钟）
      */
-    private Byte reminder;
+    private Integer reminder;
     
     /**
      * 显示颜色
